Added file for function dealing with basic API. Updated naming in db file
